feat: #31 🧪 finished tests for useBalance, ♻ some refactoring
