+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "mochaExplorer.files": "test/**/*.{j,t}s",
+  "mochaExplorer.parallel": true,
+  "mochaExplorer.timeout": 20000,
+  "terminal.integrated.fontSize": 12
+}
