+++ helpers/constants/testConstants.ts
@@ -1,0 +1,2 @@
+export const singleTimeout = 10000;
+export const basicGasPrice = 875000000;
