+++ mocharc.json
@@ -1,0 +1,4 @@
+{
+  "require": "ts-node/register/files",
+  "timeout": 20000
+}
